CodingProblems/serializers.py
from rest_framework import serializers
from .models import CodingProblem, TestCase

class TestCaseSerializer(serializers.ModelSerializer):
    class Meta:
        model = TestCase
        fields = '__all__'

class CodingProblemSerializer(serializers.ModelSerializer):
    test_cases = TestCaseSerializer(many=True, read_only=True)

    class Meta:
        model = CodingProblem
        fields = '__all__'
