requirements.txt
asgiref==3.7.2
boto3==1.28.1
botocore==1.31.1
Django==4.2.3
djangorestframework==3.14.0
jmespath==1.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
